requirements.txt
fastapi==0.124.2
sqlalchemy==2.0.45
sqlmodel==0.0.27
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.3.0
httpx==0.28.1